Remove more warnings

+++ 2D/Belyi_Sphere.cpp
@@ -7,8 +7,8 @@
 using namespace vb; using namespace std;
 
 Stream <Hypermap> triangulations (unsigned n) {
-	Cycles phic; for (unsigned i=0; i<n/3; ++i) phic.emplace_back (std::vector<unsigned> {i,i+n/3,i+2*n/3}); Permutation phi (phic);
-	int np = n/6; vector<unsigned> a (n/2 - np,2);
+	Cycles phic; for (unsigned long i=0; i<n/3; ++i) phic.emplace_back (std::vector<unsigned long> {i,i+n/3,i+2*n/3}); Permutation phi (phic);
+	int np = n/6; vector<unsigned long> a (n/2 - np,2);
 
 	return Stream<Hypermap> ([a,phi,n,np](Sink<Hypermap> & yield) {
 		std::vector<Hypermap> hs;
